requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
boto3==1.34.144
django-storages==1.14.3
python-decouple==3.8
dj-database-url==2.2.0
Pillow==10.4.0
celery==5.4.0
redis==5.0.7
django-redis==5.4.0
drf-spectacular==0.27.2
sentry-sdk==2.10.0
django-celery-beat==2.6.0